✨feat: local，globalの分割とplugin管理
