+++ vectormap_planning/local_planner/package.xml
@@ -1,16 +1,18 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>vectormap_planner</name>
+  <name>local_planner</name>
   <version>0.0.0</version>
-  <description>Vector map based global and local path planner.</description>
-  <maintainer email="s25s1045sc@chibateck.ac.jp">Kyo Yamashita</maintainer>
+  <description>Pluginlib-based local path planner server.</description>
+  <maintainer email="s21c1135sc@s.chibakoudai.jp">Kyo Yamashita</maintainer>
   <license>MIT</license>
 
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
 
+  <depend>frenet_planner</depend>
   <depend>geometry_msgs</depend>
   <depend>nav_msgs</depend>
+  <depend>pluginlib</depend>
   <depend>rclcpp</depend>
   <depend>sensor_msgs</depend>
   <depend>std_msgs</depend>
