+++ vectormap_planning/vectormap_global_planner/include/vectormap_global_planner/visibility_control.h
@@ -1,0 +1,29 @@
+#pragma once
+
+#if defined _WIN32 || defined __CYGWIN__
+  #ifdef __GNUC__
+    #define VECTORMAP_GLOBAL_PLANNER_EXPORT __attribute__ ((dllexport))
+    #define VECTORMAP_GLOBAL_PLANNER_IMPORT __attribute__ ((dllimport))
+  #else
+    #define VECTORMAP_GLOBAL_PLANNER_EXPORT __declspec(dllexport)
+    #define VECTORMAP_GLOBAL_PLANNER_IMPORT __declspec(dllimport)
+  #endif
+  #ifdef VECTORMAP_GLOBAL_PLANNER_BUILDING_LIBRARY
+    #define VECTORMAP_GLOBAL_PLANNER_PUBLIC VECTORMAP_GLOBAL_PLANNER_EXPORT
+  #else
+    #define VECTORMAP_GLOBAL_PLANNER_PUBLIC VECTORMAP_GLOBAL_PLANNER_IMPORT
+  #endif
+  #define VECTORMAP_GLOBAL_PLANNER_PUBLIC_TYPE VECTORMAP_GLOBAL_PLANNER_PUBLIC
+  #define VECTORMAP_GLOBAL_PLANNER_LOCAL
+#else
+  #define VECTORMAP_GLOBAL_PLANNER_EXPORT __attribute__ ((visibility("default")))
+  #define VECTORMAP_GLOBAL_PLANNER_IMPORT
+  #if __GNUC__ >= 4
+    #define VECTORMAP_GLOBAL_PLANNER_PUBLIC __attribute__ ((visibility("default")))
+    #define VECTORMAP_GLOBAL_PLANNER_LOCAL  __attribute__ ((visibility("hidden")))
+  #else
+    #define VECTORMAP_GLOBAL_PLANNER_PUBLIC
+    #define VECTORMAP_GLOBAL_PLANNER_LOCAL
+  #endif
+  #define VECTORMAP_GLOBAL_PLANNER_PUBLIC_TYPE
+#endif
